requirements.txt
python-dotenv
discord.py[voice]
gTTS
pubg-python